requirements.txt
Flask==0.11.1
gevent
gunicorn
keras==2.1.3
numpy
h5py
pillow
tensorflow==1.3.0